+++ test/lib/perfect_shape/test_path.rb
@@ -34,6 +34,36 @@ describe PerfectShape do
       _(shape.center_y).must_equal 150 + 35
     end
 
+    it 'constructs as closed with :wind_non_zero winding rule and shapes consisting of a nested path having lines and bezier curves not having start point (not needed)' do
+      path_shapes = []
+      path_shapes << PerfectShape::Point.new(x: 200, y: 150)
+      path_shapes << PerfectShape::Line.new(points: [270, 170])
+      path_shapes << PerfectShape::Line.new(points: [250, 220])
+      path_shapes << PerfectShape::Line.new(points: [220, 190])
+      path_shapes << PerfectShape::Line.new(points: [200, 200])
+      path_shapes << PerfectShape::Line.new(points: [180, 170])
+      path_shapes << PerfectShape::QuadraticBezierCurve.new(points: [[195, 185], [200, 190]])
+      path_shapes << PerfectShape::CubicBezierCurve.new(points: [[200, 150], [230, 160], [270, 220], [180, 170]])
+
+      path = PerfectShape::Path.new(shapes: path_shapes, closed: true, winding_rule: :wind_non_zero)
+      shape = PerfectShape::Path.new(shapes: [path], closed: true, winding_rule: :wind_non_zero)
+
+      _(shape.shapes).must_equal path_shapes
+      _(shape.points).must_equal [[200, 150], [270, 170], [250, 220], [220, 190], [200, 200], [180, 170], [195, 185], [200, 190], [200, 150], [230, 160], [270, 220], [180, 170], [200, 150]]
+      _(shape.drawing_types).must_equal [:move_to, :line_to, :line_to, :line_to, :line_to, :line_to, :quad_to, :cubic_to, :close]
+      _(shape.winding_rule).must_equal :wind_non_zero
+      _(shape.closed).must_equal true
+      _(shape.closed?).must_equal true
+      _(shape.min_x).must_equal 180
+      _(shape.min_y).must_equal 150
+      _(shape.max_x).must_equal 270
+      _(shape.max_y).must_equal 220
+      _(shape.width).must_equal 90
+      _(shape.height).must_equal 70
+      _(shape.center_x).must_equal 180 + 45
+      _(shape.center_y).must_equal 150 + 35
+    end
+
     it 'constructs as open with :wind_even_odd winding rule and shapes consisting of lines and bezier curves not having start point (not needed)' do
       path_shapes = []
       path_shapes << [200, 150]